refactor(projects): move ProjectContextPanel, ProjectChatSessionList, project_utils to sections/projects/
