+++ web/src/sections/projects/ProjectContextPanel.tsx
