+++ web/src/sections/projects/project_utils.ts
